big refactoring of Binders
