+++ hibernate-core/src/main/java/org/hibernate/boot/model/internal/MapBinder.java
@@ -53,6 +53,7 @@ import static org.hibernate.boot.model.internal.AnnotatedClassType.EMBEDDABLE;
 import static org.hibernate.boot.model.internal.AnnotatedClassType.NONE;
 import static org.hibernate.boot.model.internal.BinderHelper.findPropertyByName;
 import static org.hibernate.boot.model.internal.BinderHelper.isPrimitive;
+import static org.hibernate.boot.model.internal.EmbeddableBinder.fillEmbeddable;
 import static org.hibernate.boot.model.internal.PropertyHolderBuilder.buildPropertyHolder;
 import static org.hibernate.internal.util.StringHelper.nullIfEmpty;
 import static org.hibernate.internal.util.StringHelper.qualify;
@@ -365,7 +366,7 @@ public class MapBinder extends CollectionBinder {
 			CollectionPropertyHolder holder,
 			AccessType accessType,
 			Class<? extends CompositeUserType<?>> compositeUserType) {
-		getMap().setIndex( AnnotationBinder.fillComponent(
+		getMap().setIndex( fillEmbeddable(
 				holder,
 				propertyPreloadedData( keyClass ),
 				accessType,
