+++ hibernate-core/src/main/java/org/hibernate/boot/model/internal/PropertyBinder.java
@@ -7,64 +7,97 @@
 package org.hibernate.boot.model.internal;
 
 import java.lang.annotation.Annotation;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Member;
+import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
+import jakarta.persistence.Basic;
+import jakarta.persistence.Column;
+import jakarta.persistence.ElementCollection;
+import jakarta.persistence.EmbeddedId;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.JoinColumns;
+import jakarta.persistence.Lob;
+import jakarta.persistence.ManyToMany;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.MapsId;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.Version;
 import org.hibernate.AnnotationException;
 import org.hibernate.AssertionFailure;
-import org.hibernate.HibernateException;
 import org.hibernate.MappingException;
+import org.hibernate.annotations.Any;
 import org.hibernate.annotations.AttributeBinderType;
+import org.hibernate.annotations.CompositeType;
 import org.hibernate.annotations.IdGeneratorType;
 import org.hibernate.annotations.Immutable;
+import org.hibernate.annotations.Index;
+import org.hibernate.annotations.LazyGroup;
+import org.hibernate.annotations.ManyToAny;
 import org.hibernate.annotations.NaturalId;
 import org.hibernate.annotations.OptimisticLock;
+import org.hibernate.annotations.Parent;
 import org.hibernate.annotations.ValueGenerationType;
+import org.hibernate.annotations.common.reflection.XAnnotatedElement;
 import org.hibernate.annotations.common.reflection.XClass;
 import org.hibernate.annotations.common.reflection.XProperty;
 import org.hibernate.binder.AttributeBinder;
-import org.hibernate.boot.model.relational.ExportableProducer;
+import org.hibernate.boot.model.IdentifierGeneratorDefinition;
 import org.hibernate.boot.spi.AccessType;
 import org.hibernate.boot.spi.MetadataBuildingContext;
-import org.hibernate.generator.AnnotationBasedGenerator;
-import org.hibernate.generator.BeforeExecutionGenerator;
-import org.hibernate.generator.Generator;
-import org.hibernate.generator.GeneratorCreationContext;
-import org.hibernate.generator.OnExecutionGenerator;
-import org.hibernate.id.IdentifierGenerator;
-import org.hibernate.id.factory.spi.CustomIdGeneratorCreationContext;
+import org.hibernate.boot.spi.PropertyData;
+import org.hibernate.engine.OptimisticLockStyle;
 import org.hibernate.internal.CoreMessageLogger;
 import org.hibernate.mapping.Collection;
 import org.hibernate.mapping.Component;
 import org.hibernate.mapping.GeneratorCreator;
-import org.hibernate.mapping.IdentifierGeneratorCreator;
 import org.hibernate.mapping.KeyValue;
 import org.hibernate.mapping.MappedSuperclass;
 import org.hibernate.mapping.Property;
 import org.hibernate.mapping.RootClass;
+import org.hibernate.mapping.SimpleValue;
 import org.hibernate.mapping.ToOne;
 import org.hibernate.mapping.Value;
 import org.hibernate.metamodel.spi.EmbeddableInstantiator;
 import org.hibernate.property.access.spi.PropertyAccessStrategy;
 
+import org.hibernate.usertype.CompositeUserType;
 import org.jboss.logging.Logger;
 
-import jakarta.persistence.EmbeddedId;
-import jakarta.persistence.Id;
-import jakarta.persistence.Lob;
-import jakarta.persistence.Version;
-
-import static org.hibernate.boot.model.internal.BasicValueBinder.isOptional;
+import static jakarta.persistence.FetchType.LAZY;
+import static org.hibernate.boot.model.internal.AnyBinder.bindAny;
+import static org.hibernate.boot.model.internal.BinderHelper.isCompositeId;
+import static org.hibernate.boot.model.internal.BinderHelper.isGlobalGeneratorNameGlobal;
+import static org.hibernate.boot.model.internal.CollectionBinder.bindCollection;
+import static org.hibernate.boot.model.internal.GeneratorBinder.createForeignGenerator;
+import static org.hibernate.boot.model.internal.GeneratorBinder.createIdGenerator;
+import static org.hibernate.boot.model.internal.GeneratorBinder.generatorCreator;
 import static org.hibernate.boot.model.internal.BinderHelper.getMappedSuperclassOrNull;
+import static org.hibernate.boot.model.internal.BinderHelper.getPath;
+import static org.hibernate.boot.model.internal.BinderHelper.getPropertyOverriddenByMapperOrMapsId;
+import static org.hibernate.boot.model.internal.BinderHelper.hasToOneAnnotation;
+import static org.hibernate.boot.model.internal.GeneratorBinder.identifierGeneratorCreator;
+import static org.hibernate.boot.model.internal.GeneratorBinder.makeIdGenerator;
+import static org.hibernate.boot.model.internal.EmbeddableBinder.createCompositeBinder;
+import static org.hibernate.boot.model.internal.EmbeddableBinder.createEmbeddable;
+import static org.hibernate.boot.model.internal.EmbeddableBinder.isEmbedded;
+import static org.hibernate.boot.model.internal.HCANNHelper.findAnnotation;
 import static org.hibernate.boot.model.internal.HCANNHelper.findContainingAnnotation;
+import static org.hibernate.boot.model.internal.TimeZoneStorageHelper.resolveTimeZoneStorageCompositeUserType;
+import static org.hibernate.boot.model.internal.ToOneBinder.bindManyToOne;
+import static org.hibernate.boot.model.internal.ToOneBinder.bindOneToOne;
 import static org.hibernate.internal.util.StringHelper.qualify;
+import static org.hibernate.mapping.Constraint.hashedName;
 
 /**
+ * A stateful binder responsible for creating {@link Property} objects.
+ *
  * @author Emmanuel Bernard
  */
 public class PropertyBinder {
-	private static final CoreMessageLogger LOG = Logger.getMessageLogger(CoreMessageLogger.class, PropertyBinder.class.getName());
+	private static final CoreMessageLogger LOG = Logger.getMessageLogger( CoreMessageLogger.class, PropertyBinder.class.getName() );
 
 	private MetadataBuildingContext buildingContext;
 
@@ -84,7 +117,7 @@ public class PropertyBinder {
 	private boolean declaringClassSet;
 	private boolean embedded;
 	private EntityBinder entityBinder;
-	private boolean isXToMany;
+	private boolean toMany;
 	private String referencedEntityName;
 	private PropertyAccessStrategy propertyAccessStrategy;
 
@@ -259,34 +292,19 @@ public class PropertyBinder {
 		return bind( makePropertyAndValue() );
 	}
 
-	public void setXToMany(boolean xToMany) {
-		this.isXToMany = xToMany;
+	public void setToMany(boolean toMany) {
+		this.toMany = toMany;
 	}
 
 	private Property bind(Property property) {
 		if ( isId ) {
 			final RootClass rootClass = (RootClass) holder.getPersistentClass();
-			//if an xToMany, it has to be wrapped today.
-			//FIXME this poses a problem as the PK is the class instead of the
-			//      associated class which is not really compliant with the spec
-			if ( isXToMany || entityBinder.wrapIdsInEmbeddedComponents() ) {
-				Component identifier = (Component) rootClass.getIdentifier();
-				if (identifier == null) {
-					identifier = AnnotationBinder.createComponent(
-							holder,
-							new PropertyPreloadedData(null, null, null),
-							true,
-							false,
-							resolveCustomInstantiator(this.property, returnedClass ),
-							buildingContext
-					);
-					rootClass.setIdentifier( identifier );
-					identifier.setNullValue( "undefined" );
-					rootClass.setEmbeddedIdentifier( true );
-					rootClass.setIdentifierMapper( identifier );
-				}
+			if ( toMany || entityBinder.wrapIdsInEmbeddedComponents() ) {
+				// If an XxxToMany, it has to be wrapped today.
+				// This poses a problem as the PK is the class instead of the
+				// associated class which is not really compliant with the spec
 				//FIXME is it good enough?
-				identifier.addProperty( property );
+				getOrCreateCompositeId( rootClass ).addProperty( property );
 			}
 			else {
 				rootClass.setIdentifier( (KeyValue) getValue() );
@@ -318,17 +336,40 @@ public class PropertyBinder {
 			callAttributeBinders( property );
 		}
 		else {
-			buildingContext.getMetadataCollector().addSecondPass( persistentClasses -> callAttributeBinders( property ) );
+			buildingContext.getMetadataCollector()
+					.addSecondPass( persistentClasses -> callAttributeBinders( property ) );
 		}
 
 		return property;
+	}
+
+	private Component getOrCreateCompositeId(RootClass rootClass) {
+		final Component id = (Component) rootClass.getIdentifier();
+		if ( id == null ) {
+			final Component identifier = createEmbeddable(
+					holder,
+					new PropertyPreloadedData( null, null, null ),
+					true,
+					false,
+					resolveCustomInstantiator( property, returnedClass ),
+					buildingContext
+			);
+			rootClass.setIdentifier( identifier );
+			identifier.setNullValue( "undefined" );
+			rootClass.setEmbeddedIdentifier( true );
+			rootClass.setIdentifierMapper( identifier );
+			return identifier;
+		}
+		else {
+			return id;
+		}
 	}
 
 	private Class<? extends EmbeddableInstantiator> resolveCustomInstantiator(XProperty property, XClass embeddableClass) {
 		if ( property.isAnnotationPresent( org.hibernate.annotations.EmbeddableInstantiator.class ) ) {
 			return property.getAnnotation( org.hibernate.annotations.EmbeddableInstantiator.class ).value();
 		}
-		else if ( property.isAnnotationPresent( org.hibernate.annotations.EmbeddableInstantiator.class ) ) {
+		else if ( embeddableClass.isAnnotationPresent( org.hibernate.annotations.EmbeddableInstantiator.class ) ) {
 			return embeddableClass.getAnnotation( org.hibernate.annotations.EmbeddableInstantiator.class ).value();
 		}
 		else {
@@ -360,9 +401,29 @@ public class PropertyBinder {
 	}
 
 	private void handleValueGeneration(Property property) {
-		if ( this.property!=null ) {
+		if ( this.property != null ) {
 			property.setValueGeneratorCreator( getValueGenerationFromAnnotations( this.property ) );
 		}
+	}
+
+	/**
+	 * Returns the value generation strategy for the given property, if any.
+	 */
+	private GeneratorCreator getValueGenerationFromAnnotations(XProperty property) {
+		GeneratorCreator creator = null;
+		for ( Annotation annotation : property.getAnnotations() ) {
+			final GeneratorCreator candidate = generatorCreator( property, annotation );
+			if ( candidate != null ) {
+				if ( creator != null ) {
+					throw new AnnotationException( "Property '" + qualify( holder.getPath(), name )
+							+ "' has multiple '@ValueGenerationType' annotations" );
+				}
+				else {
+					creator = candidate;
+				}
+			}
+		}
+		return creator;
 	}
 
 	private void handleLob(Property property) {
@@ -436,172 +497,800 @@ public class PropertyBinder {
 	}
 
 	/**
-	 * Returns the value generation strategy for the given property, if any.
+	 * @param elements List of {@link PropertyData} instances
+	 * @param propertyContainer Metadata about a class and its properties
+	 *
+	 * @return the number of id properties found while iterating the elements of
+	 *         {@code annotatedClass} using the determined access strategy
 	 */
-	private GeneratorCreator getValueGenerationFromAnnotations(XProperty property) {
-		GeneratorCreator creator = null;
-		for ( Annotation annotation : property.getAnnotations() ) {
-			final GeneratorCreator candidate = generatorCreator( property, annotation );
-			if ( candidate != null ) {
-				if ( creator != null ) {
-					throw new AnnotationException( "Property '" + qualify( holder.getPath(), name )
-							+ "' has multiple '@ValueGenerationType' annotations" );
-				}
-				else {
-					creator = candidate;
-				}
+	static int addElementsOfClass(
+			List<PropertyData> elements,
+			PropertyContainer propertyContainer,
+			MetadataBuildingContext context) {
+		int idPropertyCounter = 0;
+		for ( XProperty property : propertyContainer.propertyIterator() ) {
+			idPropertyCounter += addProperty( propertyContainer, property, elements, context );
+		}
+		return idPropertyCounter;
+	}
+
+	private static int addProperty(
+			PropertyContainer propertyContainer,
+			XProperty property,
+			List<PropertyData> inFlightPropertyDataList,
+			MetadataBuildingContext context) {
+		// see if inFlightPropertyDataList already contains a PropertyData for this name,
+		// and if so, skip it..
+		for ( PropertyData propertyData : inFlightPropertyDataList ) {
+			if ( propertyData.getPropertyName().equals( property.getName() ) ) {
+				checkIdProperty( property, propertyData );
+				// EARLY EXIT!!!
+				return 0;
 			}
 		}
-		return creator;
-	}
 
-	private static void checkGeneratorClass(Class<? extends Generator> generatorClass) {
-		if ( !BeforeExecutionGenerator.class.isAssignableFrom( generatorClass )
-				&& !OnExecutionGenerator.class.isAssignableFrom( generatorClass ) ) {
-			throw new MappingException("Generator class '" + generatorClass.getName()
-					+ "' must implement either 'BeforeExecutionGenerator' or 'OnExecutionGenerator'");
-		}
-	}
+		final XClass declaringClass = propertyContainer.getDeclaringClass();
+		final XClass entity = propertyContainer.getEntityAtStake();
+		int idPropertyCounter = 0;
+		final PropertyData propertyAnnotatedElement = new PropertyInferredData(
+				declaringClass,
+				property,
+				propertyContainer.getClassLevelAccessType().getType(),
+				context.getBootstrapContext().getReflectionManager()
+		);
 
-	private static void checkGeneratorInterfaces(Class<? extends Generator> generatorClass) {
-		// we don't yet support the additional "fancy" operations of
-		// IdentifierGenerator with regular generators, though this
-		// would be extremely easy to add if anyone asks for it
-		if ( IdentifierGenerator.class.isAssignableFrom( generatorClass ) ) {
-			throw new AnnotationException("Generator class '" + generatorClass.getName()
-					+ "' implements 'IdentifierGenerator' and may not be used with '@ValueGenerationType'");
-		}
-		if ( ExportableProducer.class.isAssignableFrom( generatorClass ) ) {
-			throw new AnnotationException("Generator class '" + generatorClass.getName()
-					+ "' implements 'ExportableProducer' and may not be used with '@ValueGenerationType'");
-		}
-	}
-
-	/**
-	 * In case the given annotation is a value generator annotation, the corresponding value generation strategy to be
-	 * applied to the given property is returned, {@code null} otherwise.
-	 * Instantiates the given generator annotation type, initializing it with the given instance of the corresponding
-	 * generator annotation and the property's type.
-	 */
-	public static GeneratorCreator generatorCreator(XProperty property, Annotation annotation) {
-		final Member member = HCANNHelper.getUnderlyingMember( property );
-		final Class<? extends Annotation> annotationType = annotation.annotationType();
-		final ValueGenerationType generatorAnnotation = annotationType.getAnnotation( ValueGenerationType.class );
-		if ( generatorAnnotation == null ) {
-			return null;
-		}
-		final Class<? extends Generator> generatorClass = generatorAnnotation.generatedBy();
-		checkGeneratorClass( generatorClass );
-		checkGeneratorInterfaces( generatorClass );
-		return creationContext -> {
-			final Generator generator =
-					instantiateGenerator(
-							annotation,
-							member,
-							annotationType,
-							creationContext,
-							GeneratorCreationContext.class,
-							generatorClass
-					);
-			callInitialize( annotation, member, creationContext, generator );
-			checkVersionGenerationAlways( property, generator );
-			return generator;
-		};
-	}
-
-	public static IdentifierGeneratorCreator identifierGeneratorCreator(XProperty idProperty, Annotation annotation) {
-		final Member member = HCANNHelper.getUnderlyingMember( idProperty );
-		final Class<? extends Annotation> annotationType = annotation.annotationType();
-		final IdGeneratorType idGeneratorType = annotationType.getAnnotation( IdGeneratorType.class );
-		assert idGeneratorType != null;
-		return creationContext -> {
-			final Class<? extends Generator> generatorClass = idGeneratorType.value();
-			checkGeneratorClass( generatorClass );
-			final Generator generator =
-					instantiateGenerator(
-							annotation,
-							member,
-							annotationType,
-							creationContext,
-							CustomIdGeneratorCreationContext.class,
-							generatorClass
-					);
-			callInitialize( annotation, member, creationContext, generator );
-			checkIdGeneratorTiming( annotationType, generator );
-			return generator;
-		};
-	}
-
-	private static <C, G extends Generator> G instantiateGenerator(
-			Annotation annotation,
-			Member member,
-			Class<? extends Annotation> annotationType,
-			C creationContext,
-			Class<C> contextClass,
-			Class<? extends G> generatorClass) {
-		try {
-			try {
-				return generatorClass
-						.getConstructor( annotationType, Member.class, contextClass )
-						.newInstance( annotation, member, creationContext);
+		// put element annotated by @Id in front, since it has to be parsed
+		// before any association by Hibernate
+		final XAnnotatedElement element = propertyAnnotatedElement.getProperty();
+		if ( hasIdAnnotation( element ) ) {
+			inFlightPropertyDataList.add( 0, propertyAnnotatedElement );
+			handleIdProperty( propertyContainer, context, declaringClass, entity, element );
+			if ( hasToOneAnnotation( element ) ) {
+				context.getMetadataCollector().addToOneAndIdProperty( entity, propertyAnnotatedElement );
 			}
-			catch (NoSuchMethodException ignore) {
-				try {
-					return generatorClass
-							.getConstructor( annotationType )
-							.newInstance( annotation );
-				}
-				catch (NoSuchMethodException i) {
-					return generatorClass.newInstance();
-				}
-			}
+			idPropertyCounter++;
 		}
-		catch (InvocationTargetException|InstantiationException|IllegalAccessException e) {
-			throw new HibernateException(
-					"Could not instantiate generator of type '" + generatorClass.getName() + "'",
-					e
+		else {
+			inFlightPropertyDataList.add( propertyAnnotatedElement );
+		}
+		if ( element.isAnnotationPresent( MapsId.class ) ) {
+			context.getMetadataCollector().addPropertyAnnotatedWithMapsId( entity, propertyAnnotatedElement );
+		}
+
+		return idPropertyCounter;
+	}
+
+	private static void checkIdProperty(XProperty property, PropertyData propertyData) {
+		final Id incomingIdProperty = property.getAnnotation( Id.class );
+		final Id existingIdProperty = propertyData.getProperty().getAnnotation( Id.class );
+		if ( incomingIdProperty != null && existingIdProperty == null ) {
+			throw new MappingException(
+					String.format(
+							"You cannot override the [%s] non-identifier property from the [%s] base class or @MappedSuperclass and make it an identifier in the [%s] subclass",
+							propertyData.getProperty().getName(),
+							propertyData.getProperty().getDeclaringClass().getName(),
+							property.getDeclaringClass().getName()
+					)
 			);
 		}
 	}
 
-	private static <A extends Annotation> void callInitialize(
-			A annotation,
-			Member member,
-			GeneratorCreationContext creationContext,
-			Generator generator) {
-		if ( generator instanceof AnnotationBasedGenerator) {
-			// This will cause a CCE in case the generation type doesn't match the annotation type; As this would be
-			// a programming error of the generation type developer and thus should show up during testing, we don't
-			// check this explicitly; If required, this could be done e.g. using ClassMate
-			@SuppressWarnings("unchecked")
-			final AnnotationBasedGenerator<A> generation = (AnnotationBasedGenerator<A>) generator;
-			generation.initialize( annotation, member, creationContext );
+	private static void handleIdProperty(
+			PropertyContainer propertyContainer,
+			MetadataBuildingContext context,
+			XClass declaringClass,
+			XClass entity,
+			XAnnotatedElement element) {
+		// The property must be put in hibernate.properties as it's a system wide property. Fixable?
+		//TODO support true/false/default on the property instead of present / not present
+		//TODO is @Column mandatory?
+		//TODO add method support
+		if ( context.getBuildingOptions().isSpecjProprietarySyntaxEnabled() ) {
+			if ( element.isAnnotationPresent( Id.class ) && element.isAnnotationPresent( Column.class ) ) {
+				final String columnName = element.getAnnotation( Column.class ).name();
+				for ( XProperty property : declaringClass.getDeclaredProperties( AccessType.FIELD.getType() ) ) {
+					if ( !property.isAnnotationPresent( MapsId.class ) && isJoinColumnPresent( columnName, property ) ) {
+						//create a PropertyData for the specJ property holding the mapping
+						context.getMetadataCollector().addPropertyAnnotatedWithMapsIdSpecj(
+								entity,
+								new PropertyInferredData(
+										declaringClass,
+										//same dec
+										property,
+										// the actual @XToOne property
+										propertyContainer.getClassLevelAccessType().getType(),
+										//TODO we should get the right accessor but the same as id would do
+										context.getBootstrapContext().getReflectionManager()
+								),
+								element.toString()
+						);
+					}
+				}
+			}
 		}
 	}
 
-	private static void checkVersionGenerationAlways(XProperty property, Generator generator) {
-		if ( property.isAnnotationPresent(Version.class) ) {
-			if ( !generator.generatesOnInsert() ) {
-				throw new AnnotationException("Property '" + property.getName()
-						+ "' is annotated '@Version' but has a 'Generator' which does not generate on inserts"
+	private static boolean isJoinColumnPresent(String columnName, XProperty property) {
+		//The detection of a configured individual JoinColumn differs between Annotation
+		//and XML configuration processing.
+		if ( property.isAnnotationPresent( JoinColumn.class )
+				&& property.getAnnotation( JoinColumn.class ).name().equals( columnName ) ) {
+			return true;
+		}
+		else if ( property.isAnnotationPresent( JoinColumns.class ) ) {
+			for ( JoinColumn columnAnnotation : property.getAnnotation( JoinColumns.class ).value() ) {
+				if ( columnName.equals( columnAnnotation.name() ) ) {
+					return true;
+				}
+			}
+		}
+		return false;
+	}
+
+	private static boolean hasIdAnnotation(XAnnotatedElement element) {
+		return element.isAnnotationPresent( Id.class )
+			|| element.isAnnotationPresent( EmbeddedId.class );
+	}
+
+	/**
+	 * Process annotation of a particular property or field.
+	 */
+	public static void processElementAnnotations(
+			PropertyHolder propertyHolder,
+			Nullability nullability,
+			PropertyData inferredData,
+			Map<String, IdentifierGeneratorDefinition> classGenerators,
+			EntityBinder entityBinder,
+			boolean isIdentifierMapper,
+			boolean isComponentEmbedded,
+			boolean inSecondPass,
+			MetadataBuildingContext context,
+			Map<XClass, InheritanceState> inheritanceStatePerClass) throws MappingException {
+
+		if ( alreadyProcessedBySuper( propertyHolder, inferredData, entityBinder ) ) {
+			LOG.debugf(
+					"Skipping attribute [%s : %s] as it was already processed as part of super hierarchy",
+					inferredData.getClassOrElementName(),
+					inferredData.getPropertyName()
+			);
+		}
+		else {
+			// inSecondPass can only be used to apply right away the second pass of a composite-element
+			// Because it's a value type, there is no bidirectional association, hence second pass
+			// ordering does not matter
+
+			if ( LOG.isTraceEnabled() ) {
+				LOG.tracev(
+						"Processing annotations of {0}.{1}" ,
+						propertyHolder.getEntityName(),
+						inferredData.getPropertyName()
 				);
 			}
-			if ( !generator.generatesOnUpdate() ) {
-				throw new AnnotationException("Property '" + property.getName()
-						+ "' is annotated '@Version' but has a 'Generator' which does not generate on updates"
+
+			final XProperty property = inferredData.getProperty();
+			if ( property.isAnnotationPresent( Parent.class ) ) {
+				handleParentProperty( propertyHolder, inferredData, property );
+			}
+			else {
+				//prepare PropertyBinder
+				buildProperty(
+						propertyHolder,
+						nullability,
+						inferredData,
+						classGenerators,
+						entityBinder,
+						isIdentifierMapper,
+						isComponentEmbedded,
+						inSecondPass,
+						context,
+						inheritanceStatePerClass,
+						property,
+						inferredData.getClassOrElement()
 				);
 			}
 		}
 	}
 
-	private static void checkIdGeneratorTiming(Class<? extends Annotation> annotationType, Generator generator) {
-		if ( !generator.generatesOnInsert() ) {
-			throw new MappingException( "Annotation '" + annotationType
-					+ "' is annotated 'IdGeneratorType' but the given 'Generator' does not generate on inserts");
+	private static boolean alreadyProcessedBySuper(PropertyHolder holder, PropertyData data, EntityBinder binder) {
+		return !holder.isComponent()
+			&& binder.isPropertyDefinedInSuperHierarchy( data.getPropertyName() );
+	}
+
+	private static void handleParentProperty(PropertyHolder holder, PropertyData data, XProperty property) {
+		if ( holder.isComponent() ) {
+			holder.setParentProperty( property.getName() );
 		}
-		if ( generator.generatesOnUpdate() ) {
-			throw new MappingException( "Annotation '" + annotationType
-					+ "' is annotated 'IdGeneratorType' but the given 'Generator' generates on updates (it must generate only on inserts)");
+		else {
+			throw new AnnotationException( "Property '" + getPath( holder, data )
+					+ "' is annotated '@Parent' but is not a member of an embeddable class" );
+		}
+	}
+
+	private static void buildProperty(
+			PropertyHolder propertyHolder,
+			Nullability nullability,
+			PropertyData inferredData,
+			Map<String, IdentifierGeneratorDefinition> classGenerators,
+			EntityBinder entityBinder,
+			boolean isIdentifierMapper,
+			boolean isComponentEmbedded,
+			boolean inSecondPass,
+			MetadataBuildingContext context,
+			Map<XClass, InheritanceState> inheritanceStatePerClass,
+			XProperty property,
+			XClass returnedClass) {
+
+		final ColumnsBuilder columnsBuilder = new ColumnsBuilder(
+				propertyHolder,
+				nullability,
+				property,
+				inferredData,
+				entityBinder,
+				context
+		).extractMetadata();
+
+		final PropertyBinder propertyBinder = new PropertyBinder();
+		propertyBinder.setName( inferredData.getPropertyName() );
+		propertyBinder.setReturnedClassName( inferredData.getTypeName() );
+		propertyBinder.setAccessType( inferredData.getDefaultAccess() );
+		propertyBinder.setHolder( propertyHolder );
+		propertyBinder.setProperty( property );
+		propertyBinder.setReturnedClass( inferredData.getPropertyClass() );
+		propertyBinder.setBuildingContext( context );
+		if ( isIdentifierMapper ) {
+			propertyBinder.setInsertable( false );
+			propertyBinder.setUpdatable( false );
+		}
+		propertyBinder.setDeclaringClass( inferredData.getDeclaringClass() );
+		propertyBinder.setEntityBinder( entityBinder );
+		propertyBinder.setInheritanceStatePerClass( inheritanceStatePerClass );
+		propertyBinder.setId( !entityBinder.isIgnoreIdAnnotations() && hasIdAnnotation( property ) );
+
+		final LazyGroup lazyGroupAnnotation = property.getAnnotation( LazyGroup.class );
+		if ( lazyGroupAnnotation != null ) {
+			propertyBinder.setLazyGroup( lazyGroupAnnotation.value() );
+		}
+
+		final AnnotatedJoinColumns joinColumns = columnsBuilder.getJoinColumns();
+		final AnnotatedColumns columns = bindProperty(
+				propertyHolder,
+				nullability,
+				inferredData,
+				classGenerators,
+				entityBinder,
+				isIdentifierMapper,
+				isComponentEmbedded,
+				inSecondPass,
+				context,
+				inheritanceStatePerClass,
+				property,
+				returnedClass,
+				columnsBuilder,
+				propertyBinder
+		);
+		addIndexes( inSecondPass, property, columns, joinColumns );
+		addNaturalIds( inSecondPass, property, columns, joinColumns );
+	}
+
+	private static AnnotatedColumns bindProperty(
+			PropertyHolder propertyHolder,
+			Nullability nullability,
+			PropertyData inferredData,
+			Map<String, IdentifierGeneratorDefinition> classGenerators,
+			EntityBinder entityBinder,
+			boolean isIdentifierMapper,
+			boolean isComponentEmbedded,
+			boolean inSecondPass,
+			MetadataBuildingContext context,
+			Map<XClass, InheritanceState> inheritanceStatePerClass,
+			XProperty property,
+			XClass returnedClass,
+			ColumnsBuilder columnsBuilder,
+			PropertyBinder propertyBinder) {
+		if ( isVersion( property ) ) {
+			bindVersionProperty(
+					propertyHolder,
+					inferredData,
+					isIdentifierMapper,
+					context,
+					inheritanceStatePerClass,
+					columnsBuilder.getColumns(),
+					propertyBinder
+			);
+		}
+		else if ( isManyToOne( property ) ) {
+			bindManyToOne(
+					propertyHolder,
+					inferredData,
+					isIdentifierMapper,
+					inSecondPass,
+					context,
+					property,
+					columnsBuilder.getJoinColumns(),
+					propertyBinder,
+					isForcePersist( property )
+			);
+		}
+		else if ( isOneToOne( property ) ) {
+			bindOneToOne(
+					propertyHolder,
+					inferredData,
+					isIdentifierMapper,
+					inSecondPass,
+					context,
+					property,
+					columnsBuilder.getJoinColumns(),
+					propertyBinder,
+					isForcePersist( property )
+			);
+		}
+		else if ( isAny( property ) ) {
+			bindAny(
+					propertyHolder,
+					nullability,
+					inferredData,
+					entityBinder,
+					isIdentifierMapper,
+					context,
+					property,
+					columnsBuilder.getJoinColumns(),
+					isForcePersist( property )
+			);
+		}
+		else if ( isCollection( property ) ) {
+			bindCollection(
+					propertyHolder,
+					nullability,
+					inferredData,
+					classGenerators,
+					entityBinder,
+					isIdentifierMapper,
+					context,
+					inheritanceStatePerClass,
+					property,
+					columnsBuilder.getJoinColumns()
+			);
+		}
+		//Either a regular property or a basic @Id or @EmbeddedId while not ignoring id annotations
+		else if ( !propertyBinder.isId() || !entityBinder.isIgnoreIdAnnotations() ) {
+			// returns overridden columns
+			return bindBasic(
+					propertyHolder,
+					nullability,
+					inferredData,
+					classGenerators,
+					entityBinder,
+					isIdentifierMapper,
+					isComponentEmbedded,
+					context,
+					inheritanceStatePerClass,
+					property,
+					columnsBuilder,
+					columnsBuilder.getColumns(),
+					returnedClass,
+					propertyBinder
+			);
+		}
+		return columnsBuilder.getColumns();
+	}
+
+	private static boolean isVersion(XProperty property) {
+		return property.isAnnotationPresent( Version.class );
+	}
+
+	private static boolean isOneToOne(XProperty property) {
+		return property.isAnnotationPresent( OneToOne.class );
+	}
+
+	private static boolean isManyToOne(XProperty property) {
+		return property.isAnnotationPresent( ManyToOne.class );
+	}
+
+	private static boolean isAny(XProperty property) {
+		return property.isAnnotationPresent( Any.class );
+	}
+
+	private static boolean isCollection(XProperty property) {
+		return property.isAnnotationPresent( OneToMany.class )
+			|| property.isAnnotationPresent( ManyToMany.class )
+			|| property.isAnnotationPresent( ElementCollection.class )
+			|| property.isAnnotationPresent( ManyToAny.class );
+}
+
+	private static boolean isForcePersist(XProperty property) {
+		return property.isAnnotationPresent( MapsId.class )
+			|| property.isAnnotationPresent( Id.class );
+	}
+
+	private static void bindVersionProperty(
+			PropertyHolder propertyHolder,
+			PropertyData inferredData,
+			boolean isIdentifierMapper,
+			MetadataBuildingContext context,
+			Map<XClass, InheritanceState> inheritanceStatePerClass,
+			AnnotatedColumns columns,
+			PropertyBinder propertyBinder) {
+		checkVersionProperty( propertyHolder, isIdentifierMapper );
+		if ( LOG.isTraceEnabled() ) {
+			LOG.tracev( "{0} is a version property", inferredData.getPropertyName() );
+		}
+		final RootClass rootClass = (RootClass) propertyHolder.getPersistentClass();
+		propertyBinder.setColumns( columns );
+		final Property property = propertyBinder.makePropertyValueAndBind();
+		propertyBinder.getBasicValueBinder().setVersion( true );
+		rootClass.setVersion( property );
+
+		//If version is on a mapped superclass, update the mapping
+		final org.hibernate.mapping.MappedSuperclass superclass = getMappedSuperclassOrNull(
+				inferredData.getDeclaringClass(),
+				inheritanceStatePerClass,
+				context
+		);
+		if ( superclass != null ) {
+			superclass.setDeclaredVersion( property );
+		}
+		else {
+			//we know the property is on the actual entity
+			rootClass.setDeclaredVersion( property );
+		}
+
+		( (SimpleValue) property.getValue() ).setNullValue( "undefined" );
+		rootClass.setOptimisticLockStyle( OptimisticLockStyle.VERSION );
+		if ( LOG.isTraceEnabled() ) {
+			LOG.tracev( "Version name: {0}, unsavedValue: {1}", rootClass.getVersion().getName(),
+					( (SimpleValue) rootClass.getVersion().getValue() ).getNullValue() );
+		}
+	}
+
+	private static void checkVersionProperty(PropertyHolder propertyHolder, boolean isIdentifierMapper) {
+		if (isIdentifierMapper) {
+			throw new AnnotationException( "Class '" + propertyHolder.getEntityName()
+					+ "' is annotated '@IdClass' and may not have a property annotated '@Version'"
+			);
+		}
+		if ( !( propertyHolder.getPersistentClass() instanceof RootClass ) ) {
+			throw new AnnotationException( "Entity '" + propertyHolder.getEntityName()
+					+ "' is a subclass in an entity class hierarchy and may not have a property annotated '@Version'" );
+		}
+		if ( !propertyHolder.isEntity() ) {
+			throw new AnnotationException( "Embedded class '" + propertyHolder.getEntityName()
+					+ "' may not have a property annotated '@Version'" );
+		}
+	}
+
+	private static AnnotatedColumns bindBasic(
+			PropertyHolder propertyHolder,
+			Nullability nullability,
+			PropertyData inferredData,
+			Map<String, IdentifierGeneratorDefinition> classGenerators,
+			EntityBinder entityBinder,
+			boolean isIdentifierMapper,
+			boolean isComponentEmbedded,
+			MetadataBuildingContext context,
+			Map<XClass, InheritanceState> inheritanceStatePerClass,
+			XProperty property,
+			ColumnsBuilder columnsBuilder,
+			AnnotatedColumns columns,
+			XClass returnedClass,
+			PropertyBinder propertyBinder) {
+
+		// overrides from @MapsId or @IdClass if needed
+		final boolean isComposite;
+		final boolean isOverridden;
+		final AnnotatedColumns actualColumns;
+		if ( propertyBinder.isId() || propertyHolder.isOrWithinEmbeddedId() || propertyHolder.isInIdClass() ) {
+			// the associated entity could be using an @IdClass making the overridden property a component
+			final PropertyData overridingProperty = getPropertyOverriddenByMapperOrMapsId(
+					propertyBinder.isId(),
+					propertyHolder,
+					property.getName(),
+					context
+			);
+			if ( overridingProperty != null ) {
+				isOverridden = true;
+				final InheritanceState state = inheritanceStatePerClass.get( overridingProperty.getClassOrElement() );
+				isComposite = state != null ? state.hasIdClassOrEmbeddedId() : isEmbedded( property, returnedClass );
+				//Get the new column
+				actualColumns = columnsBuilder.overrideColumnFromMapperOrMapsIdProperty( propertyBinder.isId() );
+			}
+			else {
+				isOverridden = false;
+				isComposite = isEmbedded( property, returnedClass );
+				actualColumns = columns;
+			}
+		}
+		else {
+			isOverridden = false;
+			isComposite = isEmbedded( property, returnedClass );
+			actualColumns = columns;
+		}
+
+		final Class<? extends CompositeUserType<?>> compositeUserType =
+				resolveCompositeUserType( inferredData, context );
+
+		if ( isComposite || compositeUserType != null ) {
+			propertyBinder = createCompositeBinder(
+					propertyHolder,
+					inferredData,
+					entityBinder,
+					isIdentifierMapper,
+					isComponentEmbedded,
+					context,
+					inheritanceStatePerClass,
+					property,
+					actualColumns,
+					returnedClass,
+					propertyBinder,
+					isOverridden,
+					compositeUserType
+			);
+		}
+		else if ( property.isCollection() && property.getElementClass() != null
+				&& isEmbedded( property, property.getElementClass() ) ) {
+			// This is a special kind of basic aggregate component array type
+			// todo: see HHH-15830
+			throw new AnnotationException(
+					"Property '" + BinderHelper.getPath( propertyHolder, inferredData )
+							+ "' is mapped as basic aggregate component array, but this is not yet supported."
+			);
+		}
+		else {
+			createBasicBinder(
+					propertyHolder,
+					inferredData,
+					nullability,
+					context,
+					property,
+					actualColumns,
+					propertyBinder,
+					isOverridden
+			);
+		}
+		if ( isOverridden ) {
+			handleGeneratorsForOverriddenId(
+					propertyHolder,
+					classGenerators,
+					context,
+					property,
+					propertyBinder
+			);
+		}
+		else if ( propertyBinder.isId() ) {
+			//components and regular basic types create SimpleValue objects
+			processId(
+					propertyHolder,
+					inferredData,
+					(SimpleValue) propertyBinder.getValue(),
+					classGenerators,
+					isIdentifierMapper,
+					context
+			);
+		}
+		return actualColumns;
+	}
+
+	private static void handleGeneratorsForOverriddenId(
+			PropertyHolder propertyHolder,
+			Map<String, IdentifierGeneratorDefinition> classGenerators,
+			MetadataBuildingContext context,
+			XProperty property,
+			PropertyBinder propertyBinder) {
+		final PropertyData mapsIdProperty = getPropertyOverriddenByMapperOrMapsId(
+				propertyBinder.isId(),
+				propertyHolder,
+				property.getName(),
+				context
+		);
+		final IdentifierGeneratorDefinition foreignGenerator = createForeignGenerator( mapsIdProperty );
+		if ( isGlobalGeneratorNameGlobal( context ) ) {
+			context.getMetadataCollector()
+					.addSecondPass( new IdGeneratorResolverSecondPass(
+							(SimpleValue) propertyBinder.getValue(),
+							property,
+							foreignGenerator.getStrategy(),
+							foreignGenerator.getName(),
+							context,
+							foreignGenerator
+					) );
+		}
+		else {
+			final Map<String, IdentifierGeneratorDefinition> generators = new HashMap<>( classGenerators );
+			generators.put( foreignGenerator.getName(), foreignGenerator );
+			makeIdGenerator(
+					(SimpleValue) propertyBinder.getValue(),
+					property,
+					foreignGenerator.getStrategy(),
+					foreignGenerator.getName(),
+					context,
+					generators
+			);
+		}
+	}
+
+	private static void createBasicBinder(
+			PropertyHolder propertyHolder,
+			PropertyData inferredData,
+			Nullability nullability,
+			MetadataBuildingContext context,
+			XProperty property,
+			AnnotatedColumns columns,
+			PropertyBinder propertyBinder,
+			boolean isOverridden) {
+
+		if ( shouldForceNotNull( nullability, propertyBinder, isExplicitlyOptional( property ) ) ) {
+			forceColumnsNotNull( propertyHolder, inferredData, columns, propertyBinder );
+		}
+
+		propertyBinder.setLazy( isLazy( property ) );
+		propertyBinder.setColumns( columns );
+		if ( isOverridden ) {
+			final PropertyData mapsIdProperty = getPropertyOverriddenByMapperOrMapsId(
+					propertyBinder.isId(),
+					propertyHolder,
+					property.getName(),
+					context
+			);
+			propertyBinder.setReferencedEntityName( mapsIdProperty.getClassOrElementName() );
+		}
+
+		propertyBinder.makePropertyValueAndBind();
+	}
+
+	private static void forceColumnsNotNull(
+			PropertyHolder holder,
+			PropertyData data,
+			AnnotatedColumns columns,
+			PropertyBinder binder) {
+		for ( AnnotatedColumn column : columns.getColumns() ) {
+			if ( binder.isId() && column.isFormula() ) {
+				throw new CannotForceNonNullableException( "Identifier property '"
+						+ getPath( holder, data ) + "' cannot map to a '@Formula'" );
+			}
+			column.forceNotNull();
+		}
+	}
+
+	private static boolean shouldForceNotNull(Nullability nullability, PropertyBinder binder, boolean optional) {
+		return binder.isId()
+			|| !optional && nullability != Nullability.FORCED_NULL;
+	}
+
+	/**
+	 * Should this property be considered optional, without considering
+	 * whether it is primitive?
+	 */
+	private static boolean isExplicitlyOptional(XProperty property) {
+		return !property.isAnnotationPresent( Basic.class )
+			|| property.getAnnotation( Basic.class ).optional();
+	}
+
+	/**
+	 * Should this property be considered optional, taking into
+	 * account whether it is primitive?
+	 */
+	private static boolean isOptional(XProperty property) {
+		return property.isAnnotationPresent( Basic.class )
+				? property.getAnnotation( Basic.class ).optional()
+				: property.isArray() || !property.getClassOrElementClass().isPrimitive();
+	}
+
+	private static boolean isLazy(XProperty property) {
+		return property.isAnnotationPresent( Basic.class )
+			&& property.getAnnotation( Basic.class ).fetch() == LAZY;
+	}
+
+	private static void addIndexes(
+			boolean inSecondPass,
+			XProperty property,
+			AnnotatedColumns columns,
+			AnnotatedJoinColumns joinColumns) {
+		//process indexes after everything: in second pass, many to one has to be done before indexes
+		final Index index = property.getAnnotation( Index.class );
+		if ( index != null ) {
+			if ( joinColumns != null ) {
+				for ( AnnotatedColumn column : joinColumns.getColumns() ) {
+					column.addIndex( index, inSecondPass);
+				}
+			}
+			else {
+				if ( columns != null ) {
+					for ( AnnotatedColumn column : columns.getColumns() ) {
+						column.addIndex( index, inSecondPass );
+					}
+				}
+			}
+		}
+	}
+
+	private static void addNaturalIds(
+			boolean inSecondPass,
+			XProperty property,
+			AnnotatedColumns columns,
+			AnnotatedJoinColumns joinColumns) {
+		// Natural ID columns must reside in one single UniqueKey within the Table.
+		// For now, simply ensure consistent naming.
+		// TODO: AFAIK, there really isn't a reason for these UKs to be created
+		// on the SecondPass. This whole area should go away...
+		final NaturalId naturalId = property.getAnnotation( NaturalId.class );
+		if ( naturalId != null ) {
+			if ( joinColumns != null ) {
+				final String keyName = "UK_" + hashedName( joinColumns.getTable().getName() + "_NaturalID" );
+				for ( AnnotatedColumn column : joinColumns.getColumns() ) {
+					column.addUniqueKey( keyName, inSecondPass );
+				}
+			}
+			else {
+				final String keyName = "UK_" + hashedName( columns.getTable().getName() + "_NaturalID" );
+				for ( AnnotatedColumn column : columns.getColumns() ) {
+					column.addUniqueKey( keyName, inSecondPass );
+				}
+			}
+		}
+	}
+
+	private static Class<? extends CompositeUserType<?>> resolveCompositeUserType(
+			PropertyData inferredData,
+			MetadataBuildingContext context) {
+		final XProperty property = inferredData.getProperty();
+		final XClass returnedClass = inferredData.getClassOrElement();
+
+		if ( property != null ) {
+			final CompositeType compositeType = findAnnotation( property, CompositeType.class );
+			if ( compositeType != null ) {
+				return compositeType.value();
+			}
+			final Class<? extends CompositeUserType<?>> compositeUserType =
+					resolveTimeZoneStorageCompositeUserType( property, returnedClass, context );
+			if ( compositeUserType != null ) {
+				return compositeUserType;
+			}
+		}
+
+		if ( returnedClass != null ) {
+			final Class<?> embeddableClass = context.getBootstrapContext()
+					.getReflectionManager()
+					.toClass( returnedClass );
+			if ( embeddableClass != null ) {
+				return context.getMetadataCollector().findRegisteredCompositeUserType( embeddableClass );
+			}
+		}
+
+		return null;
+	}
+
+	private static void processId(
+			PropertyHolder propertyHolder,
+			PropertyData inferredData,
+			SimpleValue idValue,
+			Map<String, IdentifierGeneratorDefinition> classGenerators,
+			boolean isIdentifierMapper,
+			MetadataBuildingContext context) {
+		if ( isIdentifierMapper ) {
+			throw new AnnotationException( "Property '"+ getPath( propertyHolder, inferredData )
+					+ "' belongs to an '@IdClass' and may not be annotated '@Id' or '@EmbeddedId'" );
+		}
+		final XProperty idProperty = inferredData.getProperty();
+		final Annotation idGeneratorAnnotation = findContainingAnnotation( idProperty, IdGeneratorType.class );
+		final Annotation generatorAnnotation = findContainingAnnotation( idProperty, ValueGenerationType.class );
+		//TODO: validate that we don't have too many generator annotations and throw
+		if ( idGeneratorAnnotation != null ) {
+			idValue.setCustomIdGeneratorCreator( identifierGeneratorCreator( idProperty, idGeneratorAnnotation ) );
+		}
+		else if ( generatorAnnotation != null ) {
+//			idValue.setCustomGeneratorCreator( generatorCreator( idProperty, generatorAnnotation ) );
+			throw new AnnotationException( "Property '"+ getPath( propertyHolder, inferredData )
+					+ "' is annotated '" + generatorAnnotation.annotationType() + "' which is not an '@IdGeneratorType'" );
+		}
+		else {
+			final XClass entityClass = inferredData.getClassOrElement();
+			createIdGenerator( idValue, classGenerators, context, entityClass, idProperty );
+			if ( LOG.isTraceEnabled() ) {
+				LOG.tracev(
+						"Bind {0} on {1}",
+						isCompositeId( entityClass, idProperty ) ? "@EmbeddedId" : "@Id",
+						inferredData.getPropertyName()
+				);
+			}
 		}
 	}
 }
